pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "UV package managed project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.5.4",
    "fastapi>=0.115.12",
    "google-genai>=1.16.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "yfinance>=0.2.61",
]
